FEM: shell thickness and elment geometry 2D object, make name fit in FEM name system
