+++ src/Mod/Fem/App/CMakeLists.txt
@@ -97,7 +97,7 @@ SET(FemObjectsScripts_SRCS
     PyObjects/_FemMeshGroup.py
     PyObjects/_FemMeshRegion.py
     PyObjects/_FemResultMechanical.py
-    PyObjects/_FemShellThickness.py
+    PyObjects/_FemElementGeometry2D.py
     PyObjects/_FemSolverCalculix.py
     PyObjects/_FemSolverZ88.py
     PyObjects/_FemMaterial.py
@@ -122,7 +122,7 @@ SET(FemGuiScripts_SRCS
     PyGui/_CommandFemMeshRegion.py
     PyGui/_CommandFemResultShow.py
     PyGui/_CommandFemResultsPurge.py
-    PyGui/_CommandFemShellThickness.py
+    PyGui/_CommandFemElementGeometry2D.py
     PyGui/_CommandFemSolverCalculix.py
     PyGui/_CommandFemSolverControl.py
     PyGui/_CommandFemSolverRun.py
@@ -134,7 +134,7 @@ SET(FemGuiScripts_SRCS
     PyGui/_TaskPanelFemMeshGroup.py
     PyGui/_TaskPanelFemMeshRegion.py
     PyGui/_TaskPanelFemResultShow.py
-    PyGui/_TaskPanelFemShellThickness.py
+    PyGui/_TaskPanelFemElementGeometry2D.py
     PyGui/_TaskPanelFemSolverCalculix.py
     PyGui/_ViewProviderFemElementGeometry1D.py
     PyGui/_ViewProviderFemConstraintSelfWeight.py
@@ -145,7 +145,7 @@ SET(FemGuiScripts_SRCS
     PyGui/_ViewProviderFemMeshGroup.py
     PyGui/_ViewProviderFemMeshRegion.py
     PyGui/_ViewProviderFemResultMechanical.py
-    PyGui/_ViewProviderFemShellThickness.py
+    PyGui/_ViewProviderFemElementGeometry2D.py
     PyGui/_ViewProviderFemSolverCalculix.py
     PyGui/_ViewProviderFemSolverZ88.py
     PyGui/TaskPanelFemElementGeometry1D.ui
@@ -155,7 +155,7 @@ SET(FemGuiScripts_SRCS
     PyGui/TaskPanelFemMeshGroup.ui
     PyGui/TaskPanelFemMeshRegion.ui
     PyGui/TaskPanelFemResultShow.ui
-    PyGui/TaskPanelFemShellThickness.ui
+    PyGui/TaskPanelFemElementGeometry2D.ui
     PyGui/TaskPanelFemSolverCalculix.ui
 )
