+++ src/Mod/Fem/Gui/Workbench.cpp
@@ -78,7 +78,7 @@ Gui::ToolBarItem* Workbench::setupToolBars() const
          << "FEM_MaterialFluid"
          << "FEM_MaterialMechanicalNonlinear"
          << "FEM_ElementGeometry1D"
-         << "FEM_ShellThickness"
+         << "FEM_ElementGeometry2D"
          << "FEM_FluidSection"
          << "Separator"
          << "FEM_ConstraintFixed"
@@ -153,7 +153,7 @@ Gui::MenuItem* Workbench::setupMenuBar() const
          << "FEM_MaterialFluid"
          << "FEM_MaterialMechanicalNonlinear"
          << "FEM_ElementGeometry1D"
-         << "FEM_ShellThickness"
+         << "FEM_ElementGeometry2D"
          << "FEM_FluidSection"
          << "Separator"
          << "FEM_ConstraintFixed"
