+++ src/Mod/Fem/PyObjects/_FemElementGeometry2D.py
@@ -20,65 +20,21 @@
 # *                                                                         *
 # ***************************************************************************
 
-__title__ = "_ViewProviderFemShellThickness"
+__title__ = "_FemElementGeometry2D"
 __author__ = "Bernd Hahnebach"
 __url__ = "http://www.freecadweb.org"
 
-## @package ViewProviderFemShellThickness
+## @package FemElementGeometry2D
 #  \ingroup FEM
 
-import FreeCAD
-import FreeCADGui
-from pivy import coin
 
+class _FemElementGeometry2D:
+    "The FemElementGeometry2D object"
+    def __init__(self, obj):
+        obj.addProperty("App::PropertyLength", "Thickness", "ShellThickness", "set thickness of the shell elements")
+        obj.addProperty("App::PropertyLinkSubList", "References", "ShellThickness", "List of shell thickness shapes")
+        obj.Proxy = self
+        self.Type = "FemElementGeometry2D"
 
-class _ViewProviderFemShellThickness:
-    "A View Provider for the FemShellThickness object"
-    def __init__(self, vobj):
-        vobj.Proxy = self
-
-    def getIcon(self):
-        return ":/icons/fem-shell-thickness.svg"
-
-    def attach(self, vobj):
-        self.ViewObject = vobj
-        self.Object = vobj.Object
-        self.standard = coin.SoGroup()
-        vobj.addDisplayMode(self.standard, "Standard")
-
-    def getDisplayModes(self, obj):
-        return ["Standard"]
-
-    def getDefaultDisplayMode(self):
-        return "Standard"
-
-    def updateData(self, obj, prop):
+    def execute(self, obj):
         return
-
-    def onChanged(self, vobj, prop):
-        return
-
-    def setEdit(self, vobj, mode=0):
-        import PyGui._TaskPanelFemShellThickness
-        taskd = PyGui._TaskPanelFemShellThickness._TaskPanelFemShellThickness(self.Object)
-        taskd.obj = vobj.Object
-        FreeCADGui.Control.showDialog(taskd)
-        return True
-
-    def unsetEdit(self, vobj, mode=0):
-        FreeCADGui.Control.closeDialog()
-        return
-
-    def doubleClicked(self, vobj):
-        doc = FreeCADGui.getDocument(vobj.Object.Document)
-        if not doc.getInEdit():
-            doc.setEdit(vobj.Object.Name)
-        else:
-            FreeCAD.Console.PrintError('Active Task Dialog found! Please close this one first!\n')
-        return True
-
-    def __getstate__(self):
-        return None
-
-    def __setstate__(self, state):
-        return None
